requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
openai
hypothesis
pytest
pyyaml
httpx
